requirements.txt
phidata
openai
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
groq
uvicorn